WebApp/requirements.txt
bcrypt==4.1.2
cryptography==42.0.2
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
Jinja2==3.1.3
numpy==1.24.4
pillow==10.2.0
PyJWT==2.8.0
pytest==8.0.0
pytz==2024.1
requests==2.31.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0